add loading state, changes logo, and responsive
