+++ src/pages/Register.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import axios from "axios";
 import { Link } from "react-router-dom";
-import { BiLogInCircle } from "react-icons/bi";
 import GoogleLogin from "../components/GoogleLogin";
+import { BiSolidUserDetail } from "react-icons/bi"
 
 const Register = () => {
   const [firstname, setFirstName] = useState("");
@@ -80,7 +80,7 @@ const Register = () => {
       <div className="absolute top-28 p-10 w-full lg:top-14 lg:w-full lg:px-[350px]">
         <div className="border border-red-600 rounded-xl p-5">
           <div className="text-3xl font-bold mb-4 flex flex-row items-center justify-center text-red-600">
-            <BiLogInCircle className="text-5xl" />
+            <BiSolidUserDetail className="text-6xl" />
             <h1>Register</h1>
           </div>
           <form
